Basic chat implementation
This is not yet a real time chat implementation, but it works.

The chat creator is an actor that can create new chats.

A chat is an entity that offers the affordance of creating a participation.

A participation is like an invite link that's associated with a name.

Anyone who has a participation URL can submit messages under that name.

I've also changed some various things in the actor framework.

It's still a work in progress and this is the first real feature
that involves multiple participants and so on.

The next thing to add is real-time updates which I'm still thinking about.

I'll write about that in a note.